requirements.txt
Flask==3.0.3
itsdangerous==2.2.0
PyYAML==6.0.1
requests==2.31.0
Werkzeug==3.0.3